requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
streamlit>=1.30.0
pypdf2>=3.0.0
python-docx>=1.1.0
reportlab>=4.0.0